+++ src/PDFBuilder.js
@@ -169,7 +169,9 @@ export default class PDFBuilder extends WebformBuilder {
     // Normal PDF Builder
     return super.attach(element).then(() => {
       this.loadRefs(this.element, {
-        iframeDropzone: 'single', 'sidebar-container': 'multiple'
+        iframeDropzone: 'single',
+        'sidebar-container': 'multiple',
+        'sidebar-loader': 'single',
       });
 
       this.afterAttach();
@@ -184,6 +186,11 @@ export default class PDFBuilder extends WebformBuilder {
     this.on('removeComponent', (component) => {
       this.webform.postMessage({ name: 'removeElement', data: component });
     });
+    if (this.refs['sidebar-loader']) {
+      this.webform.on('iframe-ready', () => {
+        this.refs['sidebar-loader'].remove();
+      }, true);
+    }
     this.initIframeEvents();
     this.updateDropzoneDimensions();
     this.initDropzoneEvents();
@@ -228,6 +235,7 @@ export default class PDFBuilder extends WebformBuilder {
   createForm(options) {
     // Instantiate the webform from the PDF class instead of Webform
     options.skipInit = false;
+    options.hideLoader = true;
     this.webform = new PDF(this.element, options);
     this.webform.on('attach', () => {
       // If the dropzone exists but has been removed in a PDF rebuild, reinstate it
@@ -260,7 +268,7 @@ export default class PDFBuilder extends WebformBuilder {
         originalComponent = comp;
         return true;
       }
-    });
+    }, true);
     return {
       formioComponent: component.parent,
       formioContainer: container,
@@ -275,6 +283,7 @@ export default class PDFBuilder extends WebformBuilder {
     this.webform.on('iframe-elementUpdate', schema => {
       const component = this.webform.getComponentById(schema.id);
       if (component && component.component) {
+        const isNew = true;
         component.component.overlay = {
           page: schema.page,
           left: schema.left,
@@ -284,7 +293,7 @@ export default class PDFBuilder extends WebformBuilder {
         };
 
         if (!this.options.noNewEdit && !component.component.noNewEdit) {
-          this.editComponent(component.component, this.getParentContainer(component));
+          this.editComponent(component.component, this.getParentContainer(component), isNew);
         }
         this.emit('updateComponent', component.component);
       }
@@ -340,6 +349,11 @@ export default class PDFBuilder extends WebformBuilder {
     });
 
     this.addEventListener(this.refs.iframeDropzone, 'drop', this.onDropzoneDrop.bind(this));
+  }
+
+  updateDragAndDrop() {
+    this.initDropzoneEvents();
+    this.prepSidebarComponentsForDrag();
   }
 
   prepSidebarComponentsForDrag() {
@@ -416,7 +430,7 @@ export default class PDFBuilder extends WebformBuilder {
 
     // Set a unique key for this component.
     BuilderUtils.uniquify([this.webform.component], schema);
-    this.webform.component.components.push(schema);
+    this.webform._form.components.push(schema);
 
     schema.overlay = {
       top: layerY - this.itemOffsetY + HEIGHT,