+++ src/components/_classes/nestedarray/NestedArrayComponent.js
@@ -28,17 +28,38 @@ export default class NestedArrayComponent extends NestedDataComponent {
     this._rowIndex = value;
   }
 
+  init() {
+    super.init();
+    this.prevHasAddButton = this.hasAddButton();
+  }
+
+  checkAddButtonChanged() {
+    const isAddButton = this.hasAddButton();
+    if (isAddButton !== this.prevHasAddButton) {
+      this.prevHasAddButton = isAddButton;
+      this.redraw();
+    }
+  }
+
   checkData(data, flags, row) {
     data = data || this.rootValue;
     flags = flags || {};
     row = row || this.data;
+    this.checkAddButtonChanged();
 
     return this.checkRows('checkData', data, flags, Component.prototype.checkData.call(this, data, flags, row));
   }
 
   checkRows(method, data, opts, defaultValue, silentCheck) {
     return this.iteratableRows.reduce(
-      (valid, row) => this.checkRow(method, data, opts, row.data, row.components, silentCheck) && valid,
+      (valid, row, rowIndex) => {
+        if (!opts?.rowIndex || opts?.rowIndex === rowIndex) {
+          return this.checkRow(method, data, opts, row.data, row.components, silentCheck) && valid;
+        }
+        else {
+          return valid;
+        }
+      },
       defaultValue,
     );
   }
@@ -92,7 +113,7 @@ export default class NestedArrayComponent extends NestedDataComponent {
         else if (fn) {
           fn(component, components);
         }
-        result = rowIndex !== null ? comp : result.concat(comp);
+        result = rowIndex !== null ? comp : result.concat(comp || possibleComp);
       }
     }, rowIndex);
     if ((!result || result.length === 0) && possibleComp) {
@@ -133,7 +154,7 @@ export default class NestedArrayComponent extends NestedDataComponent {
             <tr>
       `);
 
-      this.component.components.forEach((component) => {
+      this.component.components?.forEach((component) => {
         const label = component.label || component.key;
         result += `<th style="padding: 5px 10px;">${label}</th>`;
       });
@@ -162,6 +183,10 @@ export default class NestedArrayComponent extends NestedDataComponent {
       `);
 
       return result;
+    }
+
+    if (!value || !value.length) {
+      return '';
     }
 
     return super.getValueAsString(value, options);