+++ src/components/datagrid/DataGrid.js
@@ -38,7 +38,11 @@ export default class DataGridComponent extends NestedArrayComponent {
 
     // Add new values based on minLength.
     this.rows = [];
-    this.createRows(true);
+
+    if (this.initRows) {
+      this.createRows(true);
+    }
+
     this.visibleColumns = {};
     this.checkColumns();
   }
@@ -57,6 +61,14 @@ export default class DataGridComponent extends NestedArrayComponent {
 
   get defaultSchema() {
     return DataGridComponent.schema();
+  }
+
+  get initEmpty() {
+    return this.component.initEmpty || this.component.noFirstRow;
+  }
+
+  get initRows() {
+    return this.builderMode || this.path === 'defaultValue' || !this.initEmpty;
   }
 
   get emptyValue() {
@@ -139,7 +151,7 @@ export default class DataGridComponent extends NestedArrayComponent {
   getRowChunks(groups, rows) {
     const [, chunks] = groups.reduce(
       ([startIndex, acc], size) => {
-        const endIndex = startIndex +  size;
+        const endIndex = startIndex + size;
         return [endIndex, [...acc, [startIndex, endIndex]]];
       }, [0, []]
     );
@@ -296,7 +308,7 @@ export default class DataGridComponent extends NestedArrayComponent {
       this.refs[`${this.datagridKey}-group-header`].forEach((header, index) => {
         this.addEventListener(header, 'click', () => this.toggleGroup(header, index));
       });
-     }
+    }
 
     const columns = this.getColumns();
     const rowLength = columns.length;
@@ -347,6 +359,7 @@ export default class DataGridComponent extends NestedArrayComponent {
 
     this.rows[index] = this.createRowComponents(this.dataValue[index], index);
     this.checkConditions();
+    this.triggerChange();
     this.redraw();
   }
 
@@ -354,6 +367,7 @@ export default class DataGridComponent extends NestedArrayComponent {
     this.splice(index);
     const [row] = this.rows.splice(index, 1);
     _.each(row, (component) => this.removeComponent(component));
+    this.setValue(this.dataValue, { isReordered: true });
     this.redraw();
   }
 
@@ -394,6 +408,9 @@ export default class DataGridComponent extends NestedArrayComponent {
       const options = _.clone(this.options);
       options.name += `[${rowIndex}]`;
       options.row = `${rowIndex}-${colIndex}`;
+      if (col.id) {
+        col.id = col.id + rowIndex;
+      }
       const component = this.createComponent(col, options, row);
       component.parentDisabled = !!this.disabled;
       component.rowIndex = rowIndex;
@@ -492,13 +509,15 @@ export default class DataGridComponent extends NestedArrayComponent {
 
     // Make sure we always have at least one row.
     // NOTE: Removing this will break "Public Configurations" in portal. ;)
-    if (value && !value.length && !this.component.noFirstRow) {
+    if (value && !value.length && !this.initEmpty) {
       value.push({});
     }
 
     const changed = this.hasChanged(value, this.dataValue);
-    this.dataValue = value;
-    this.createRows();
+    if (this.initRows) {
+      this.dataValue = value;
+      this.createRows();
+    }
     this.rows.forEach((row, rowIndex) => {
       if (value.length <= rowIndex) {
         return;
@@ -524,6 +543,17 @@ export default class DataGridComponent extends NestedArrayComponent {
     if (_.isNumber(key) && remainingPath.length) {
       const compKey = remainingPath.pop();
       result = this.rows[key][compKey];
+      // If the component is inside a Layout Component, try to find it among all the row's components
+      if (!result) {
+        Object.entries(this.rows[key]).forEach(([, comp]) => {
+          if ('getComponent' in comp) {
+            const possibleResult = comp.getComponent([compKey], fn);
+            if (possibleResult) {
+              result = possibleResult;
+            }
+          }
+        });
+      }
       if (result && _.isFunction(fn)) {
         fn(result, this.getComponents());
       }