+++ src/components/editgrid/EditGrid.js
@@ -10,6 +10,7 @@ const EditRowState = {
   New: 'new',
   Editing: 'editing',
   Saved: 'saved',
+  Viewing: 'viewing',
   Removed: 'removed',
   Draft: 'draft',
 };
@@ -50,34 +51,38 @@ export default class EditGridComponent extends NestedArrayComponent {
 
   static get defaultHeaderTemplate() {
     return `<div class="row">
-  {% (component.components || []).forEach(function(component) { %}
-    <div class="col-sm-2">{{ component.label }}</div>
-  {% }) %}
-</div>`;
+      {% util.eachComponent(components, function(component) { %}
+        {% if (!component.hasOwnProperty('tableView') || component.tableView) { %}
+          <div class="col-sm-2">{{ component.label }}</div>
+        {% } %}
+      {% }) %}
+    </div>`;
   }
 
   static get defaultRowTemplate() {
     return `<div class="row">
-  {% instance.eachComponent(function(component) { %}
-    <div class="col-sm-2">
-      {{ component.getView(component.dataValue) }}
-    </div>
-  {% }, rowIndex) %}
-  {% if (!instance.options.readOnly && !instance.originalComponent.disabled) { %}
-    <div class="col-sm-2">
-      <div class="btn-group pull-right">
-        <button class="btn btn-default btn-light btn-sm editRow"><i class="{{ iconClass('edit') }}"></i></button>
-        {% if (!instance.hasRemoveButtons || instance.hasRemoveButtons()) { %}
-          <button class="btn btn-danger btn-sm removeRow"><i class="{{ iconClass('trash') }}"></i></button>
+      {% util.eachComponent(components, function(component) { %}
+        {% if (!component.hasOwnProperty('tableView') || component.tableView) { %}
+          <div class="col-sm-2">
+            {{ getView(component, row[component.key]) }}
+          </div>
         {% } %}
-      </div>
-    </div>
-  {% } %}
-</div>`;
+      {% }) %}
+      {% if (!instance.disabled) { %}
+        <div class="col-sm-2">
+          <div class="btn-group pull-right">
+            <button class="btn btn-default btn-light btn-sm editRow"><i class="{{ iconClass('edit') }}"></i></button>
+            {% if (!instance.hasRemoveButtons || instance.hasRemoveButtons()) { %}
+              <button class="btn btn-danger btn-sm removeRow"><i class="{{ iconClass('trash') }}"></i></button>
+            {% } %}
+          </div>
+        </div>
+      {% } %}
+    </div>`;
   }
 
   get defaultDialogTemplate() {
-    return  `
+    return `
     <h3 ref="dialogHeader">${this.t('Do you want to clear data?')}</h3>
     <div style="display:flex; justify-content: flex-end;">
       <button ref="dialogCancelButton" class="btn btn-secondary">${this.t('Cancel')}</button>
@@ -184,8 +189,8 @@ export default class EditGridComponent extends NestedArrayComponent {
 
     const massageContainerRef = 'messageContainer';
 
-    if (refs[`${ massageContainerRef }`] === 'single') {
-      this.refs[`${ massageContainerRef }`] = element.querySelector(`:scope > [ref="${ massageContainerRef }"]`);
+    if (refs[`${massageContainerRef}`] === 'single') {
+      this.refs[`${massageContainerRef}`] = [...element.children].filter(elem => elem.attributes?.ref?.value === massageContainerRef)[0];
     }
   }
 
@@ -235,7 +240,7 @@ export default class EditGridComponent extends NestedArrayComponent {
   }
 
   isOpen(editRow) {
-    return [EditRowState.New, EditRowState.Editing].includes(editRow.state);
+    return [EditRowState.New, EditRowState.Editing, EditRowState.Viewing].includes(editRow.state);
   }
 
   render(children) {
@@ -305,12 +310,13 @@ export default class EditGridComponent extends NestedArrayComponent {
             className: 'editRow',
             event: 'click',
             action: () => {
-              this.editRow(rowIndex).then(()=> {
+              this.editRow(rowIndex).then(() => {
                 if (this.component.rowDrafts) {
                   this.validateRow(editRow, false);
 
-                  if (this.component.modal && editRow.errors && !!editRow.errors.length ) {
+                  if (this.component.modal && editRow.errors && !!editRow.errors.length) {
                     this.alert.showErrors(editRow.errors, false);
+                    editRow.alerts = true;
                   }
                 }
               });
@@ -376,7 +382,7 @@ export default class EditGridComponent extends NestedArrayComponent {
             const instance = flattenedComponents[component.key];
             let view = instance ? instance.getView(data || instance.dataValue) : '';
 
-            if (instance && instance.widget && (view !== '--- PROTECTED ---' )) {
+            if (instance && instance.widget && (view !== '--- PROTECTED ---')) {
               if (_.isArray(view)) {
                 view = view.map((value) => instance.widget.getValueAsString(value));
               }
@@ -469,7 +475,7 @@ export default class EditGridComponent extends NestedArrayComponent {
   }
 
   addRowModal(rowIndex) {
-    const modalContent =  this.ce('div');
+    const modalContent = this.ce('div');
     const editRow = this.editRows[rowIndex];
     editRow.willBeSaved = false;
     const { components } = editRow;
@@ -501,6 +507,7 @@ export default class EditGridComponent extends NestedArrayComponent {
         }
         else {
           this.alert.showErrors(editRow.errors, false);
+          editRow.alerts = true;
         }
       },
     }, this.component.saveRow || 'Save'));
@@ -515,7 +522,7 @@ export default class EditGridComponent extends NestedArrayComponent {
     }
 
     const wrapper = this.ce('div', { ref: 'confirmationDialog' });
-    const dialogContent =this.component.dialogTemplate || this.defaultDialogTemplate;
+    const dialogContent = this.component.dialogTemplate || this.defaultDialogTemplate;
 
     wrapper.innerHTML = dialogContent;
     wrapper.refs = {};
@@ -554,10 +561,7 @@ export default class EditGridComponent extends NestedArrayComponent {
       return;
     }
     editRow.prevState = editRow.state;
-
-    if (!this.options.readOnly) {
-      editRow.state = EditRowState.Editing;
-    }
+    editRow.state = this.options.readOnly ? EditRowState.Viewing : EditRowState.Editing;
 
     const dataSnapshot = fastCloneDeep(editRow.data);
 
@@ -674,6 +678,10 @@ export default class EditGridComponent extends NestedArrayComponent {
     this.checkValidity(null, true);
     this.redraw();
 
+    if (editRow.alerts) {
+      editRow.alerts = false;
+    }
+
     return true;
   }
 
@@ -697,7 +705,7 @@ export default class EditGridComponent extends NestedArrayComponent {
 
   updateRowsComponents(rowIndex) {
     this.editRows.slice(rowIndex).forEach((row, index) => {
-      this.updateComponentsRowIndex(row.components, index);
+      this.updateComponentsRowIndex(row.components, rowIndex + index);
     });
   }
 
@@ -736,6 +744,13 @@ export default class EditGridComponent extends NestedArrayComponent {
         const editRow = this.editRows[rowIndex];
         if (this.inlineEditMode) {
           this.triggerRootChange(flags, changed, modified);
+        }
+        else if (editRow?.alerts) {
+          this.checkData(null, {
+            ...flags,
+            changed,
+            rowIndex,
+          }, this.data);
         }
         else if (editRow) {
           this.checkRow('checkData', null, {
@@ -790,7 +805,23 @@ export default class EditGridComponent extends NestedArrayComponent {
 
     editRow.errors = !valid ? this.errors.filter((err) => !errorsSnapshot.includes(err)) : null;
 
+    this.showRowErrorAlerts(editRow, !!valid);
+
     return !!valid;
+  }
+
+  showRowErrorAlerts(editRow, valid) {
+    if (editRow.alerts) {
+      if (this.alert) {
+        if (editRow.errors?.length && !valid) {
+          this.alert.showErrors(editRow.errors, false);
+          editRow.alerts = true;
+        }
+        else {
+          this.alert.clear();
+        }
+      }
+    }
   }
 
   checkValidity(data, dirty, row) {
@@ -819,7 +850,7 @@ export default class EditGridComponent extends NestedArrayComponent {
 
     this.editRows.forEach((editRow, index) => {
       // Trigger all errors on the row.
-      const rowValid = this.validateRow(editRow, dirty);
+      const rowValid = this.validateRow(editRow, editRow.alerts || dirty);
 
       rowsValid &= rowValid;
 
@@ -831,7 +862,7 @@ export default class EditGridComponent extends NestedArrayComponent {
         if (rowContainer) {
           const errorContainer = rowContainer.querySelector('.editgrid-row-error');
 
-          if (!rowValid ) {
+          if (!rowValid) {
             errorContainer.textContent = 'Invalid row. Please correct it or delete.';
           }
         }