+++ src/components/number/Number.js
@@ -177,7 +177,7 @@ export default class NumberComponent extends Input {
   focus() {
     const input = this.refs.input[0];
     if (input) {
-      input.focus();
+      super.focus.call(this);
       input.setSelectionRange(0, input.value.length);
     }
   }