+++ src/components/signature/Signature.js
@@ -81,7 +81,7 @@ export default class SignatureComponent extends Input {
     }
 
     if (this.signaturePad && this.dataValue && this.signaturePad.isEmpty()) {
-      this.signaturePad.fromDataURL(this.dataValue);
+      this.setDataToSigaturePad();
     }
 
     return changed;
@@ -141,7 +141,7 @@ export default class SignatureComponent extends Input {
       this.signaturePad.clear();
 
       if (this.dataValue) {
-        this.signaturePad.fromDataURL(this.dataValue);
+        this.setDataToSigaturePad();
       }
     }
   }
@@ -230,5 +230,13 @@ export default class SignatureComponent extends Input {
 
   focus() {
     this.refs.padBody.focus();
+  }
+
+  setDataToSigaturePad() {
+    this.signaturePad.fromDataURL(this.dataValue, {
+      ratio: 1,
+      width: this.refs.canvas.width,
+      height: this.refs.canvas.height,
+    });
   }
 }