+++ src/components/well/editForm/Well.edit.display.js
@@ -27,4 +27,8 @@ export default [
     key: 'tabindex',
     ignore: true
   },
+  {
+    key: 'tableView',
+    ignore: true
+  },
 ];